refactor: change mass variable from float to double

+++ src/main/java/be/archilios/experiment/engine/DrawableItem.java
@@ -9,7 +9,7 @@ public abstract class DrawableItem extends Item implements Drawable {
         super(location);
     }
     
-    public DrawableItem(Coordinate2D location, float mass) {
+    public DrawableItem(Coordinate2D location, double mass) {
         super(location, mass);
     }
 }
